refactor: reservationslot 으로 이름 변경
