+++ backend/src/main/java/org/example/backend/reservationslot/exception/ReservationException.java
@@ -1,4 +1,4 @@
-package org.example.backend.reservation.exception;
+package org.example.backend.reservationslot.exception;
 
 import lombok.RequiredArgsConstructor;
 import org.example.backend.common.exception.BaseException;
