+++ backend/src/main/java/org/example/backend/reservationslot/exception/ReservationExceptionType.java
@@ -1,4 +1,4 @@
-package org.example.backend.reservation.exception;
+package org.example.backend.reservationslot.exception;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
