+++ backend/src/main/java/org/example/backend/reservationslot/repository/ReservationSlotRepository.java
@@ -1,13 +1,13 @@
-package org.example.backend.reservation.repository;
+package org.example.backend.reservationslot.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import org.example.backend.reservation.domain.Reservation;
+import org.example.backend.reservationslot.domain.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationSlotRepository extends JpaRepository<Reservation, Long> {
     @Query("SELECT COUNT(r) > 0 FROM Reservation r " +
             "WHERE r.room.id = :seminarRoomId " +
             "AND r.startTime < :endTime " +
